exception handling, password encoding

+++ src/main/java/com/quynhtien/cineasy/service/UserService.java
@@ -4,12 +4,15 @@ import com.quynhtien.cineasy.dto.request.UserCreationRequest;
 import com.quynhtien.cineasy.dto.request.UserUpdateRequest;
 import com.quynhtien.cineasy.dto.response.UserResponse;
 import com.quynhtien.cineasy.entity.User;
+import com.quynhtien.cineasy.exception.AppException;
+import com.quynhtien.cineasy.exception.ErrorCode;
 import com.quynhtien.cineasy.mapper.UserMapper;
 import com.quynhtien.cineasy.repository.UserRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,6 +24,7 @@ import java.util.List;
 public class UserService {
     UserRepository userRepository;
     UserMapper userMapper;
+    PasswordEncoder passwordEncoder;
 
     //Get all users
     public List<UserResponse> getUsers() {
@@ -31,17 +35,16 @@ public class UserService {
     //Get user by id
     public UserResponse getUser(String id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
         return userMapper.toUserResponse(user);
     }
 
     //Create user
     public UserResponse createUser(UserCreationRequest request) {
         User user = userMapper.toUser(request);
-        log.info("request {}", request.getFirstName());
-        log.info("user {}", user.getFirstName());
+
         //encode password
-        //
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
 
         userRepository.save(user);
         return userMapper.toUserResponse(user);
@@ -50,8 +53,10 @@ public class UserService {
     //Update user
     public UserResponse updateUser(String id, UserUpdateRequest request) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+
         //encode password
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
 
         userMapper.updateUser(request, user);
         userRepository.save(user);
@@ -60,6 +65,9 @@ public class UserService {
 
     //Delete user
     public String deleteUser(String id) {
+        if (!userRepository.existsById(id)) {
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.deleteById(id);
         return "Delete user successfully";
     }
